Allow smaller hashes than curve sizes in PKI signs
In #14994, it was noticed that larger hashes than curve sizes could be
used for the majority of algorithms that it shouldn't be allowed for.
E.g., with P-256, a 512-bit SHA-2 invocation could be used, resulting in
an incorrect prefix truncation (different from the standardized
SHA-2-512/256 by NIST). This is non-standard and non-desirable.

However, in attempting to fix it, we broke users who were specifying
smaller hash functions than curve sizes. While this makes a little less
sense in general (why not use a smaller curve?) since support seems to
be fairly widespread for this, we attempt the following behavior:

 - If signature_bits is not specified, size based off the curve size.
 - If signature_bits is too large, size based off the curve size.
 - If signature_bits is the right size or smaller, use it.

This should allow for a reasonable rolling default with the value of
key_bits=0, while allowing undersized hashes when necessary.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>